ddb/pyproject.toml
[project]
name = "darbotdb"
version = "0.4.0"
description = "DarbotDB (DDB) — composable agent database with graph storage, SQLite micro DBs, AG-UI sessions, and MCP apps"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "python-arango>=8.0.0",  # DarbotDB graph driver (fork-compatible)
  "orjson>=3.10",
  "aiosqlite>=0.20",
  "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
]

[project.scripts]
darbotdb-api = "darbotdb.api.main:run"
ddb-api = "darbotdb.api.main:run"

[tool.setuptools]
packages = ["darbotdb", "darango", "cards", "graph", "graph3d", "micro", "session", "triad", "agui", "txt2kg"]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["."]
